@@ -3,44 +3,44 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/material": "^5.2.7",
-    "@mui/styled-engine-sc": "^5.1.0",
-    "@mui/x-data-grid": "^5.2.2",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.24.0",
-    "dotenv": "^12.0.3",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/material": "^5.6.2",
+    "@mui/styled-engine-sc": "^5.6.1",
+    "@mui/x-data-grid": "^5.9.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "axios": "^0.26.1",
+    "dotenv": "^16.0.0",
     "emailjs-com": "^3.2.0",
-    "firebase": "^9.6.9",
-    "flag": "^4.4.0",
+    "firebase": "^9.6.11",
+    "flag": "^5.0.1",
     "jwt-decode": "^3.1.2",
     "khalti-checkout-web": "^2.2.0",
-    "material-ui-chip-input": "^2.0.0-beta.2",
+    "material-ui-chip-input": "1.1.0",
     "moment": "^2.29.1",
-    "notistack": "^2.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "notistack": "^2.0.4",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-file-base64": "^1.0.3",
     "react-google-login": "^5.2.2",
     "react-icons": "^4.3.1",
-    "react-redux": "^7.2.6",
+    "react-redux": "^8.0.1",
     "react-reveal": "^1.2.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
-    "react-scroll": "^1.8.4",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "react-scroll": "^1.8.7",
     "react-swipeable-views": "^0.14.0",
-    "redux": "^4.1.2",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.4.1",
-    "styled-components": "^5.3.3",
-    "web-vitals": "^2.1.2"
+    "styled-components": "^5.3.5",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
