@@ -1,7 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
+import { legacy_createStore as createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { reducers } from './redux/reducers/index'
 import App from './App';
@@ -9,10 +9,10 @@ import './index.css';
 import { composeWithDevTools } from 'redux-devtools-extension'
 const store = createStore(reducers, composeWithDevTools(applyMiddleware(thunk)));
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <Provider store={store}>
     <App />
   </Provider>,
-  document.getElementById('root')
 );
 
